+++ aidial_sdk/_pydantic/__init__.py
@@ -1,0 +1,2 @@
+from ._compat import PYDANTIC_V2, ConfigDict
+from ._models import BaseModel
