+++ tests/test_serialization.py
@@ -1,11 +1,8 @@
 import json
 
-from aidial_sdk._pydantic._compat import (
-    model_dump,
-)
 from aidial_sdk.chat_completion import Message, ResponseFormatJsonSchema, Role
 from aidial_sdk.chat_completion.request import ResponseFormatJsonSchemaObject
-from tests.utils.pydantic import model_parse, model_parse_json
+from tests.utils.pydantic import model_dump, model_parse, model_parse_json
 
 
 def test_message_ser():
