+++ tests/utils/errors.py
@@ -1,6 +1,6 @@
 from pydantic import BaseModel
 
-from aidial_sdk._pydantic._compat import PYDANTIC_V2
+from aidial_sdk._pydantic import PYDANTIC_V2
 
 
 class Error(BaseModel):
